+++ CHANGELOG.md
@@ -13,7 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Emit `.align 3` directives for strings for `EEGCC`.
+- Emit `.align 3` directives for strings for `EEGCC` compiler.
+- Emit `.align 3` directives for functions for `EEGCC` compiler.
 
 ## [1.23.1] - 2024-03-22